# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/spaces/availability": {
      "get": {
        "operationId": "SpacesController_checkAvailability",
        "summary": "Cek Ketersediaan Space Berdasarkan Tanggal & Jam",
        "parameters": [
          {
            "name": "id_space",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "tanggal",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-08-30",
              "type": "string"
            }
          },
          {
            "name": "jam_mulai",
            "required": true,
            "in": "query",
            "schema": {
              "example": "09:00",
              "type": "string"
            }
          },
          {
            "name": "durasi_jam",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Spaces (Katalog & Ketersediaan)"
        ]
      }
    },
    "/api/spaces": {
      "get": {
        "operationId": "SpacesController_getAll",
        "summary": "Lihat Semua Space Coworking (Filter ?tipe & ?search)",
        "parameters": [
          {
            "name": "tipe",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "desk",
                "meeting_room",
                "private_office"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Spaces (Katalog & Ketersediaan)"
        ]
      }
    },
    "/api/spaces/{id}": {
      "get": {
        "operationId": "SpacesController_getById",
        "summary": "Lihat Detail Space Coworking Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Spaces (Katalog & Ketersediaan)"
        ]
      }
    },
...
    "/api/admin/spaces/{id}": {
      "get": {
        "operationId": "AdminController_getSpaceById",
        "summary": "Detail Data Space Berdasarkan ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "AdminController_updateSpace",
        "summary": "Update Data Ruangan & Fasilitas Space",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteSpace",
        "summary": "Hapus Data Ruangan / Meja Space",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin Panel"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Smart Space Booking API",
    "description": "Backend RESTful API untuk Sistem Reservasi Coworking Space - UKK RPL 2026/2027 SMK Telkom Malang",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateSpaceDto": {
        "type": "object",
        "properties": {
          "nama_space": {
            "type": "string",
            "example": "Personal Desk Alpha 01 (Updated)"
          },
          "harga_per_jam": {
            "type": "number",
            "example": 30000
          },
          "tipe": {
            "type": "string",
            "enum": [
              "desk",
              "meeting_room",
              "private_office"
            ]
          },
          "kapasitas": {
            "type": "number",
            "example": 2
          },
          "deskripsi": {
            "type": "string"
          },
          "foto": {
            "type": "string"
          }
        }
      },
...
    }
  }
}